services/python-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.6
alembic==1.14.1
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
pytest==8.3.5
httpx==0.28.1
